requirements.txt
Flask==3.0.0
Pillow==10.1.0
eel==0.18.1
Werkzeug==3.0.1 